requirements.txt
# Core dependencies
smolagents>=1.0.0
together>=0.2.5
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0

# Google Sheets integration
gspread>=5.12.0
oauth2client>=4.1.3

# UI
gradio>=4.0.0

# HTTP client
requests>=2.31.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0 